Toy with expressing more of the build pipeline as generators.
It might gain us a little RAM (and cache hit ratio), but my guess is only a little. We still have to concretize refs and regions eventually (and at the same time), to sort them. Likewise, annotations gets zipped and thus concretized, because the file.html template loops through the lines of the file 3 times at present. So, links and various kinds of needles are left as the possible wins. We might get a little boost from not having them concretized in RAM at the same time as refs, regions, etc.

It's also about 10 lines longer. I think the overall quality comes out about even—maybe even a little worse—and don't think it's worth debugging.